requirements.txt
flask
requests
pytest
sesamutils
pysmb
lxml
xmltodict